Feature: Vault Secrets infra + UI draft
This branch introduces the implementation of secrets for the vault domain. Secrets are sensitive data directly associated with a specific snggle object, such as "mnemonic" for Vault and "private key" for Wallet. Secrets should be securely stored in the database and only accessible when needed (for example: when signing transaction).

List of changes:
- created vault_list_item_cubit.dart and updated vault_list_item.dart to present the basic functionalities of vaults encrypting and decrypting (UI draft)
- updated vault_list_page_cubit.dart to add/delete vault secrets while creating new vault
- created secrets_repository.dart to manage CRUD operations on "secrets" database key
- created vault_secrets_service.dart to manage CRUD operations on secrets related with vault domain
- created vault_secrets_model.dart storing decrypted secrets for vault
- added new "secrets" key to database_parent_key.dart