Refactor: MasterKeyController from AuthSingletonCubit
The goal of this branch is to replace auth_singleton_cubit.dart with master_key_controller.dart. The role of Bloc/Cubit class is to manage views in the application. Previously auth_singleton_cubit.dart was responsible only for storing password for the master key and encrypting/decrypting values at different levels of the application. For example the data access layer, when values stored in database are encrypted/decrypted, or the presentation layer, when the application asks for the master key/app password or decides which view should be displayed. The previous name could be misleading so decided to use more generic name: "controller".

List of changes:
- removed auth_singleton_cubit.dart and created master_key_controller.dart which reproduces previous functionality of auth_singleton_cubit.dart