Bugfix: Camera permissions
The purpose of this branch is fix the looping camera permission prompt, that was displayed again and again if the user denied camera permissions, on Scan QR Page. This was caused by lack of custom permission handling in the app and was solved with the help of the permission_handler package.

List of changes:
- added permission_handler dependency to pubspec.yaml to allow custom permission handling
- created qr_camera_scaffold_cubit.dart to allow passing permission statuses. The current permission can be accessed globally, however reading the status is an asynchronous operation.
- modified qr_camera_scaffold.dart to display custom dialog in case of "permanently denied" status. It can only be changed to "granted" by redirecting to the app settings.