Feature: Password management
This branch introduces PasswordController, which changes the way passwords are passed in the application. Previously, passwords were passed through the constructors of individual widgets. After the changes, passwords are passed via the global PasswordController class. Adding and deleting passwords is done according to the navigation of the application. This solution allows keeping track of the application navigation, gives more control over the deletion of password data, and will allow to store passwords for a certain period of time in the future.
One of the reasons for creating this branch was the lack of a way to verify if a wallet with encrypted parents already had its path unlocked. This problem has also been solved on this branch.

List of changes:
- created password_controller.dart to store the passwords of the accessed elements globally, registered in locator.dart
- created list_item_access_model.dart to group the PasswordModel with AListItemModel, to simplify storing the records in PasswordController
- added method "getEncryptedPath()" in secrets_service.dart to allow getting information about SignTxPageCubit, which will compare the encrypted parents with the passwords stored in PasswordController
- modified "_navigateToNextPage()" method in each list page to add the passwords to PasswordController on each navigation
- modified "navigateBack()" method in a_list_cubit.dart, to delete the passwords from PasswordController on each navigation out
- replaced passwords passed by constructors with PasswordController calls, throughout the app
- modified sign_tx_page_cubit.dart to include the new criteria whether the QR scanning should be blocked