Feature: Migration to Gradle 7.6.1 (#38)
This branch focuses on upgrading Flutter's Gradle plugins versions and the way these plugins are applied. Previously, Gradle was applied using the legacy imperative apply script method. In Flutter 3.16, support was added for applying these plugins with Gradle's declarative plugins {} block (also called the Plugin DSL) and it is now the recommended approach.

List of changes:
- updated Gradle version to 7.6.1, Android Gradle Plugin (AGP) to 7.4.2 and Kotlin to 1.8.22 - these versions are compatible with Java SDK 17, which is recommended for building Flutter Android apps since Flutter 3.10
- changed Gradle plugin applying way from imperative to declarative, which is recommended for Flutter versions from 3.16