Bugfix: Text scaling
This branch fixes issues with text scaling and alignment in ShuffleButton, WalletCreatePage, WalletListItem, and related components. Ensures consistent text appearance, proper single-line rendering, and improves UI spacing for a better user experience.

List of changes:
- modified shuffle_button.dart to keep ShuffleButton text on a single line
- modified wallet_create_page.dart to remove unnecessary padding before the address index in the derivation path
- modified wallet_create_page.dart, wallet_list_item.dart and horizontal_list_item_layout.dart to correctly display the derivation path
- modified custom_dialog_option.dart to properly display option buttons

Side note: For now, we have decided not to change the widgets listed below, because most of them require a complete overhaul and will be corrected separately:
- vault_create_page.dart,
- transaction_details_page.dart,
- sign_tx_page.dart,
- wallet_connect_page.dart,
- settings_page.dart