Bugfix: Derivation path disappear (#42)
This branch fixes the issue with disappearing part of the derivation path (non-editable part), that occured on Create Wallet view when the user deleted the path index (editable part). The problem stems from how "prefixText" works as a property of InputDecoration used in TextField. This widget was designed that if the editable text is empty, the prefixText will also disappear. However, there is a popular workaround that includes another field "prefixIcon". This element does not disappear if the editable text is empty and is of type Widget, not Icon, so it can be used as Text.

List of changes:
- changed "prefixIcon" to "prefixedWidget" and "prefixIconConstraints" to "prefixWidgetConstraints" in custom_text_field.dart and added comment that explains this workaround
- replaced the "prefixText" and "prefixedStyle" usage with "prefixWidget" in wallet_create_page.dart and changed "prefixWidgetConstraints" value to properly display the TextWidget